requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
gunicorn==22.0.0
google-generativeai==0.8.3
